+++ extras/integration_tests/flutter_db/lib/moor_flutter.dart
@@ -35,18 +35,15 @@ Future<void> main() async {
 
   // Additional integration test for flutter: Test loading a database from asset
   test('can load a database from asset', () async {
-    const dbNameInDevice = 'app_from_asset.db';
-
-    final folder = await getDatabasesPath();
-    final file = File(join(folder, dbNameInDevice));
-
-    if (await file.exists()) {
-      await file.delete();
+    final databasesPath = await getDatabasesPath();
+    final dbFile = File(join(databasesPath, 'app_from_asset.db'));
+    if (await dbFile.exists()) {
+      await dbFile.delete();
     }
 
     var didCallCreator = false;
-    final executor = FlutterQueryExecutor.inDatabaseFolder(
-      path: dbNameInDevice,
+    final executor = FlutterQueryExecutor(
+      path: dbFile.path,
       singleInstance: true,
       creator: (file) async {
         final content = await rootBundle.load('test_asset.db');