+++ moor/lib/src/runtime/isolate/client.dart
@@ -139,7 +139,7 @@ class _TransactionIsolateExecutor extends _BaseExecutor
 
   Future<bool> _openAtServer() async {
     _executorId =
-        await client._channel.request(_NoArgsRequest.startTransaction) as int;
+        await client._channel.request<int>(_NoArgsRequest.startTransaction);
     return true;
   }